app/requirements.txt
Flask
mysql-connector
PyYAML
#cryptography
